requirements.txt
streamlit~=1.37.1
pandas~=2.2.3
matplotlib~=3.10.0
prophet
faiss-cpu
numpy~=1.26.4
python-dotenv~=1.1.1
sentence-transformers
plotly
google-generativeai

gitdb~=4.0.12
pathlib~=1.0.1
typing_extensions~=4.12.2
defusedxml~=0.7.1
Pillow~=9.5.0
olefile~=0.47
pip~=24.0
emoji~=2.14.0
attrs~=24.3.0
distro~=1.9.0
cryptography~=44.0.0
filelock~=3.16.1
rsa~=4.9
pyasn1~=0.6.1
pytz~=2024.2
toml~=0.10.2
requests~=2.32.5
rich~=13.9.4
keras~=3.2.1
colorama~=0.4.6
PyYAML~=6.0
torch~=2.4.1
scipy~=1.15.1
setuptools~=75.8.2
packaging~=23.2
smmap~=5.0.2
cffi~=1.17.1
numba~=0.60.0
regex~=2024.11.6
threadpoolctl~=3.5.0
mpmath~=1.3.0
sympy~=1.13.1
tabulate~=0.9.0
astunparse~=1.6.3
onnx~=1.17.0
onnxruntime~=1.17.0
tensorboard~=2.16.2
optree~=0.11.0
tqdm~=4.67.1
tensorflow~=2.16.1
torchvision~=0.20.1
psutil~=6.1.1
networkx~=3.4.2
urllib3~=2.3.0
fsspec~=2024.12.0
altair~=4.2.2
jsonschema~=4.23.0
narwhals~=1.21.1
pyarrow~=18.1.0
Jinja2~=3.1.5
referencing~=0.35.1
click~=8.1.8
aiohttp~=3.11.11
yarl~=1.18.3
pyasn1_modules~=0.4.1
cachetools~=5.5.0
certifi~=2024.12.14
google-api-python-client~=2.162.0
pydantic~=1.10.21
httplib2~=0.22.0
MarkupSafe~=3.0.2
joblib~=1.4.2
SQLAlchemy~=1.4.54
lxml~=5.3.0
numexpr~=2.10.2
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
tzdata~=2024.2
scikit-learn~=1.6.1
tornado~=6.4.2
h2~=3.2.0
six~=1.17.0
pyparsing~=3.2.1
pydantic_core~=2.23.4
Deprecated~=1.2.15
idna~=2.10
chardet~=3.0.4
tenacity~=8.5.0
watchdog~=4.0.2
contourpy~=1.3.1
fonttools~=4.55.3
blinker~=1.9.0
langchain~=0.0.184
jsonpointer~=3.0.0
cycler~=0.12.1
kiwisolver~=1.4.8
wheel~=0.43.0
Pygments~=2.19.1
tokenizers~=0.21.0
safetensors~=0.5.2
uritemplate~=4.1.1
transformers~=4.48.0
huggingface-hub~=0.34.4
timm~=1.0.13
nltk~=3.8.1
beautifulsoup4~=4.12.3
pytesseract~=0.3.13
h5py~=3.11.0
groq~=0.15.0